+++ examples/nfz-quasar-unocss-pinia-starter/nuxt.config.ts
@@ -129,6 +129,7 @@ export default defineNuxtConfig({
 
   runtimeConfig: {
     demo: {
+      enabled: process.env.NFZ_DEMO_ENABLED,
       user: process.env.NFZ_DEMO_USER || 'admin',
       password: process.env.NFZ_DEMO_PASSWORD || 'admin123',
       roles: process.env.NFZ_DEMO_ROLES || 'admin,user',